rt(vk): use manuallydrop for VulkanBuffer rather than Option
